requirements.txt
matplotlib~=3.4.2
numpy~=1.20.3
opencv-python~=4.5.3.56
torchvision~=0.10.0
torch~=1.9.0
albumentations~=1.0.3
pillow~=8.3.1